+++ libde265/sps.h
@@ -233,8 +233,8 @@ public:
 
   uint8_t WpOffsetBdShiftY;
   uint8_t WpOffsetBdShiftC;
-  int32_t WpOffsetHalfRangeY;
-  int32_t WpOffsetHalfRangeC;
+  uint16_t WpOffsetHalfRangeY;
+  uint16_t WpOffsetHalfRangeC;
 
 
   int getPUIndexRS(int pixelX,int pixelY) const {
